Rewrite everything from scratch, taking 'Emmet Cheat Sheet' as a guide
I've rewrote everything in the framework from scratch, taking
'Emmet Cheat Sheet' as a guide http://docs.emmet.io/cheat-sheet,
to preserve the similarity between abbreviations that are used in
the x-framework and those that are used in Emmet.

I've also added the first group or category of the css rules that
exist in Emmet Cheat Sheet, which called 'Visual Formatting'.
And I've made some useful functions (mixins) to simplify the work
that have been made and that will be made.